% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_initialize_gxe.R
\name{init_setup_gxe}
\alias{init_setup_gxe}
\title{Initialize a susie object}
\usage{
init_setup_gxe(
  n,
  p,
  L,
  scaled_prior_variance,
  residual_variance,
  prior_weights,
  null_weight,
  varY,
  standardize
)
}
\description{
Initialize a susie object
}
