% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_effect_regression_ss_gxe.R
\name{single_effect_regression_ss_gxe}
\alias{single_effect_regression_ss_gxe}
\title{Single effect regression}
\usage{
single_effect_regression_ss_gxe(
  Xty,
  XtX_inv,
  V = matrix(c(1, 0.5, 0.5, 1), ncol = 2),
  residual_variance = 1,
  prior_weights = NULL,
  optimize_V = c("none", "optim", "uniroot", "EM", "simple"),
  check_null_threshold = 0
)
}
\arguments{
\item{Xty}{A 2p-vector.}

\item{XtX_inv}{a return object of function S_inverse_crossprod}
}
\description{
Single effect regression
}
\keyword{internal}
