% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils_gxe.R
\name{susie_get_cs}
\alias{susie_get_cs}
\title{Get credible sets}
\usage{
susie_get_cs(
  res,
  X = NULL,
  Xcorr = NULL,
  coverage = 0.95,
  min_abs_corr = 0.5,
  dedup = TRUE,
  squared = FALSE,
  check_symmetric = TRUE,
  n_purity = 100,
  use_rfast
)
}
\arguments{
\item{X}{n by p matrix of values of the p variables (covariates) in
n samples. When provided, correlation between variables will be
computed and used to remove CSs whose minimum correlation among
variables is smaller than \code{min_abs_corr}.}

\item{Xcorr}{p by p matrix of correlations between variables
(covariates). When provided, it will be used to remove CSs whose
minimum correlation among variables is smaller than
\code{min_abs_corr}.}

\item{coverage}{A number between 0 and 1 specifying desired
coverage of each CS.}

\item{min_abs_corr}{A "purity" threshold for the CS. Any CS that
contains a pair of variables with correlation less than this
threshold will be filtered out and not reported.}

\item{dedup}{If \code{dedup = TRUE}, remove duplicate CSs.}

\item{squared}{If \code{squared = TRUE}, report min, mean and
median of squared correlation instead of the absolute correlation.}

\item{check_symmetric}{If \code{check_symmetric = TRUE}, perform a
check for symmetry of matrix \code{Xcorr} when \code{Xcorr} is
provided (not \code{NULL}).}

\item{n_purity}{The maximum number of credible set (CS) variables
used in calculating the correlation (\dQuote{purity})
statistics. When the number of variables included in the CS is
greater than this number, the CS variables are randomly subsampled.}

\item{use_rfast}{Use the Rfast package for the purity calculations.
By default \code{use_rfast = TRUE} if the Rfast package is
installed.}
}
\description{
Get credible sets
}
