% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils_gxe.R
\name{susie_get_pip}
\alias{susie_get_pip}
\title{Get PIPs}
\usage{
susie_get_pip(res, prune_by_cs = FALSE, prior_tol = 1e-09)
}
\arguments{
\item{prune_by_cs}{Whether or not to ignore single effects not in
a reported CS when calculating PIP.}

\item{prior_tol}{Filter out effects having estimated prior variance
smaller than this threshold.}
}
\description{
Get PIPs
}
