% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_rss_gxe.R
\name{susie_rss_gxe}
\alias{susie_rss_gxe}
\title{Sum of Single Effects (SuSiE) Regression using Summary Statistics from Gene by Environment (GxE) Interaction}
\usage{
susie_rss_gxe(
  z,
  R,
  n,
  bhat,
  bhat_gxe,
  shat,
  shat_gxe,
  covhat = NULL,
  var_y,
  z_ld_weight = 0,
  estimate_residual_variance = FALSE,
  prior_variance = matrix(c(50, 1, 1, 50), ncol = 2),
  check_prior = TRUE,
  ...
)
}
\arguments{
\item{z}{p-vector of z-scores. Not supported in GxE.}

\item{R}{p x p correlation matrix.}

\item{n}{The sample size.}

\item{bhat}{Alternative summary data giving the estimated effects
(a vector of length p). This, together with \code{shat},\code{shat_gxe},\code{bhat_gxe}, may be
provided instead of \code{z}.}

\item{bhat_gxe}{Alternative summary data giving the estimated effects
(a vector of length p). This, together with \code{bhat}, \code{shat},\code{shat_gxe}, \code{covhat}, may be
provided instead of \code{z}.}

\item{shat}{Alternative summary data giving the standard errors of
the estimated effects (a vector of length p). This, together with
\code{bhat}, \code{bhat_gxe},\code{shat_gxe}, \code{covhat}, may be provided instead of \code{z}.}

\item{shat_gxe}{Alternative summary data giving the standard errors of
the estimated effects (a vector of length p). This, together with
\code{bhat}, \code{shat},\code{shat_gxe}, \code{covhat}, may be provided instead of \code{z}.}

\item{covhat}{Alternative summary data giving the covariance of the estimated effects
(a vector of length p). This, together with
\code{bhat} and \code{shat}, \code{bhat_gxe},\code{shat_gxe}, may be provided instead of \code{z}.}

\item{var_y}{The sample variance of y, defined as \eqn{y'y/(n-1)}.
When the sample variance is not provided, the coefficients
(returned from \code{coef}) are computed on the
\dQuote{standardized} X, y scale. Not supported in GxE.}

\item{z_ld_weight}{This parameter is included for backwards
compatibility with previous versions of the function, but it is no
longer recommended to set this to a non-zero value. When
\code{z_ld_weight > 0}, the matrix \code{R} is adjusted to be
\code{cov2cor((1-w)*R + w*tcrossprod(z))}, where \code{w =
z_ld_weight}.}

\item{estimate_residual_variance}{The default is FALSE, the
residual variance is fixed to 1 or variance of y.}

\item{prior_variance}{The prior variance(s) for the non-zero
noncentrality parameterss \eqn{\tilde{b}_l}. It is either a 2x2 matrix (prior variance of gene main effect, gene by E and prior covariance of gene and gene by E interaction),
or a list of length L and each element is a 2x2 matrix . When the \code{susie_suff_stat_gxe} option
\code{estimate_prior_variance} is set to \code{TRUE} this simply provides an initial value for the
prior variance.}

\item{check_prior}{When \code{check_prior = TRUE}, it checks if the
estimated prior variance becomes unreasonably large (comparing with
100 * max(abs(chi-square))^2).}

\item{\dots}{Other parameters to be passed to
\code{\link{susie_suff_stat_gxe}}.}
}
\value{
A \code{"susie"} object with the following
  elements:

\item{alpha}{An L by p matrix of posterior inclusion probabilites.}

\item{mu}{An L by 2p matrix of posterior means, conditional on
  inclusion.}

\item{mu2}{A list of L 2p by 2p matrices of posterior second moments,
  conditional on inclusion.}

\item{lbf}{log-Bayes Factor for each single effect.}

\item{lbf_variable}{log-Bayes Factor for each variable and single effect.}

\item{V}{Prior variance of the non-zero elements of b.}

\item{elbo}{The value of the variational lower bound, or
  \dQuote{ELBO} (objective function to be maximized), achieved at
  each iteration of the IBSS fitting procedure.}

\item{sets}{Credible sets estimated from model fit; see
  \code{\link{susie_get_cs}} for details.}

\item{pip}{A vector of length p giving the (marginal) posterior
  inclusion probabilities for all p covariates.}

\item{niter}{Number of IBSS iterations that were performed.}

\item{converged}{\code{TRUE} or \code{FALSE} indicating whether
  the IBSS converged to a solution within the chosen tolerance
  level.}
}
\description{
\code{susie_rss_gxe} performs variable selection under a
  sparse Bayesian multiple linear regression of \eqn{Y} on \eqn{X}
  using the effect sizes estimates and SE from standard univariate regression
  of \eqn{Y} on each column of \eqn{X}, an estimate, \eqn{R}, of
  the correlation matrix for the columns of \eqn{X} (\dQuote{LD matrix}). See
  \dQuote{Details} for other ways to call \code{susie_rss}
}
\details{
With the inputs \code{bhat}, \code{bhat_gxe}, \code{shat}, \code{shat_gxe}, and \code{R},
\code{susie_rss_gxe} calls \code{susie_suff_stat_gxe} with constructed \code{XtX}, \code{XtZ}, \code{ZtZ}
and \code{Xty = dS_inv$Kty}, and with \code{residual_variance = 1}. The
underlying assumption of performing the analysis in this way is
that the sample size is large (\emph{i.e.}, infinity), and/or the
effects are small.

The \code{estimate_residual_variance} setting is \code{FALSE} by
default, which is recommended when the LD matrix is estimated from
a reference panel.
}
\references{
G. Wang, A. Sarkar, P. Carbonetto and M. Stephens (2020). A simple
new approach to variable selection in regression, with application
to genetic fine-mapping. \emph{Journal of the Royal Statistical
Society, Series B} \bold{82}, 1273-1300 \doi{10.1101/501114}.

Y. Zou, P. Carbonetto, G. Wang, G and M. Stephens
(2022). Fine-mapping from summary data with the \dQuote{Sum of
Single Effects} model. \emph{PLoS Genetics} \bold{18},
e1010299. \doi{10.1371/journal.pgen.1010299}.
}
